Fix a few MSVC warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk@27622 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/port/cpl_vsil.cpp
@@ -926,7 +926,7 @@ int VSIIngestFile( VSILFILE* fp,
                         VSIFCloseL( fp );
                     return FALSE;
                 }
-                GByte* pabyNew = (GByte*)VSIRealloc(*ppabyRet, nDataAlloc);
+                GByte* pabyNew = (GByte*)VSIRealloc(*ppabyRet, (size_t)nDataAlloc);
                 if( pabyNew == NULL )
                 {
                     CPLError( CE_Failure, CPLE_OutOfMemory,
